GENIO: usb: typec: altmodes/displayport: Fix potential overrun in pin_assignment_show
Fix coverity reported error:
"Overrunning array "pin_assignments" of 6 8-byte elements at element
index 255" in function pin_assignment_show().

This overrun does not actually happen because "assignment" is u8,
so it is not possible that

    for (i = 0; assignments; assignments >>=1, ++i)

reaches (i == 255) condition.

However, to prevent future assignment expansions, it makes sense
to properly store the result of "get_count_order" in signed int.

drivers/usb/typec/altmodes/displayport.c

(1) Event return_constant:
    Function call "get_count_order((dp->data.conf & 0xff00UL) >> 8)"
    may return -1.
(2) Event assignment:	Assigning: "cur" =
    "get_count_order((dp->data.conf & 0xff00UL) >> 8)".
    The value of "cur" is now 255

    cur = get_count_order(DP_CONF_GET_PIN_ASSIGN(dp->data.conf));

(15) Event cond_const:	Checking "i == cur" implies that "i" is
    255 on the true branch.
(16) Event overrun-local:
    Overrunning array "pin_assignments" of 6 8-byte elements at
    element index 255 (byte offset 2047)
    using index "i" (which evaluates to 255).

+++ drivers/usb/typec/altmodes/displayport.c
@@ -494,7 +494,7 @@ static ssize_t pin_assignment_show(struct device *dev,
 	struct dp_altmode *dp = dev_get_drvdata(dev);
 	u8 assignments;
 	int len = 0;
-	u8 cur;
+	int cur;
 	int i;
 
 	mutex_lock(&dp->lock);
